Added weapons menù, fixed stamina bar, added wand attack animation
Deleted the folder with the prefabs of useless weapons and wands that i had put in my previous commit.

The fireball should be released a moment later, otherwise the animation of the attack comes late and uncoordinated

+++ Unity Builds/VGD - Utilities/Assets/Scripts/Player/PlayerManager.cs
@@ -8,7 +8,7 @@ public class PlayerManager : MonoBehaviour
     public static readonly int MaxLives = 5;
     public static readonly int MaxHealth = 100;
     public static readonly int MaxMana = 100;
-    public static readonly int MaxStamina = 100;
+    public static readonly float MaxStamina = 100;
     private static readonly int DefaultLives = 3;
     private static readonly int DefaultHealth = 80;
     private static readonly int DefaultMana = 80;
@@ -23,7 +23,7 @@ public class PlayerManager : MonoBehaviour
     public static int CurrentLives;
     public static int CurrentHealth;
     public static int CurrentMana;
-    public static int CurrentStamina;
+    public static float CurrentStamina;
     
     // Start is called before the first frame update
     private void Awake()
